Remove redundant GAMS test driver
