+++ pyomo/solvers/tests/checks/test_GAMS.py
@@ -9,8 +9,10 @@
 #  This software is distributed under the 3-clause BSD License.
 #  ___________________________________________________________________________
 
-
-from pyomo.environ import ConcreteModel, Var, Objective, Constraint, maximize, Expression, log10
+import pyomo.environ as pyo
+from pyomo.environ import (
+    ConcreteModel, Var, Objective, Constraint, maximize, Expression, log10,
+)
 from pyomo.opt import SolverFactory, TerminationCondition
 
 from pyomo.solvers.plugins.solvers.GAMS import (
@@ -317,6 +319,27 @@ class GAMSTests(unittest.TestCase):
     def test_version_gms(self):
         with SolverFactory("gams", solver_io="gms") as opt:
             self.assertIsNotNone(opt.version())
+
+    @unittest.skipIf(not gamsgms_available,
+                     "The 'gams' executable is not available")
+    def test_dat_parser(self):
+        # This tests issue 2571
+        m = pyo.ConcreteModel()
+        m.S = pyo.Set(initialize=list(range(5)))
+        m.a_long_var_name = pyo.Var(m.S, bounds=(0, 1), initialize=1)
+        m.obj = pyo.Objective(
+            expr=2000 * pyo.summation(m.a_long_var_name), sense=pyo.maximize)
+        solver = pyo.SolverFactory("gams:conopt")
+        res = solver.solve(
+            m, symbolic_solver_labels=True, load_solutions=False,
+            io_options={'put_results_format': 'dat'})
+        self.assertEqual(res.solution[0].Objective['obj']['Value'], 10000)
+        for i in range(5):
+            self.assertEqual(
+                res.solution[0].Variable[f'a_long_var_name_{i}_']['Value'],
+                1
+            )
+
 
 class GAMSLogfileTestBase(unittest.TestCase):
     def setUp(self):
